feat(ref-klv-04): mirror ST 0903 u64 widening into JVM + Python bindings
JNI: add checked_u64 (rejects negative i64, passes any non-negative as u64
since Java has no unsigned primitive), swap checked_u32 -> checked_u64 for
targetId + centroidPixel + bboxTopLeft/BottomRight + centroidPixRow/Col on
decode; encode uses target_id as i64 (i64::from(u64) doesn't compile).
Python: extract::<u32> -> extract::<u64> for the same 5+1 fields in klv.rs;
errors.rs VTargetPackEmpty/DuplicateTargetId cast u64 as u32 for the
informational .tag attribute (full info in message). Java record was already
long/Long; .pyi already int — no surface change, no C ABI change. Large-value
round-trip tests on both: V6 fields (target_id + Tags 1/2/3) tested above
u32::MAX; V4 fields (Tags 19/20) capped at u32::MAX per spec §10.2.2.20/.21.

+++ bindings/python/src/errors.rs
@@ -672,8 +672,8 @@ pub(crate) fn klv_encode_error_to_pyerr(py: Python<'_>, e: tst_core::KlvEncodeEr
             ("MISSING_MANDATORY_ITEM", Some(u32::from(*tag)))
         }
         RustE::ReservedTagInUnknown { tag } => ("RESERVED_TAG_IN_UNKNOWN", Some(*tag)),
-        RustE::VTargetPackEmpty { target_id } => ("VTARGET_PACK_EMPTY", Some(*target_id)),
-        RustE::DuplicateTargetId { target_id } => ("DUPLICATE_TARGET_ID", Some(*target_id)),
+        RustE::VTargetPackEmpty { target_id } => ("VTARGET_PACK_EMPTY", Some(*target_id as u32)),
+        RustE::DuplicateTargetId { target_id } => ("DUPLICATE_TARGET_ID", Some(*target_id as u32)),
         RustE::ForbiddenStandaloneOffset { tag } => ("FORBIDDEN_STANDALONE_OFFSET", Some(*tag)),
         _ => ("BUFFER_TOO_SMALL", None),
     };
